// sync-react-temp/src/lib/Upload/types.ts
export type IConnections = {
    url: string;
    headers?: any;
}
// ...
export const imageState = {
    None: 0,
    Loading: 1,
    Done: 2,
    Problem: 3
}
// ...
export interface IDownloadFileProps {
    connection: IConnections;
    file: {
        name: string | null;
        url: string | null;
    };
    onLoading?: (value: boolean) => void;
    onError?: (status: number, data: any) => void;
    imageStatus?: (value: number) => void;
    isAborted?: boolean;
}